cases fixes

+++ libft/ft_strjoin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strjoin.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wasmar <wasmar@student.42.fr>              +#+  +:+       +#+        */
+/*   By: schaaban <schaaban@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/11 19:42:25 by wasmar            #+#    #+#             */
-/*   Updated: 2025/04/02 08:16:35 by wasmar           ###   ########.fr       */
+/*   Updated: 2025/05/13 10:45:41 by schaaban         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strjoin(char const *s1, char const *s2, t_malloc *mallo)
 {
-	size_t	i;
-	size_t	j;
+	int	i;
+	int	j;
 	char	*a;
 
 	i = 0;
