+++ libft/libft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   libft.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: wasmar <wasmar@student.42.fr>              +#+  +:+       +#+        */
+/*   By: schaaban <schaaban@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/19 14:58:10 by wasmar            #+#    #+#             */
-/*   Updated: 2025/04/02 08:16:19 by wasmar           ###   ########.fr       */
+/*   Updated: 2025/05/13 10:45:45 by schaaban         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ typedef struct s_list
 	struct s_list					*next;
 }									t_list;
 
-size_t								ft_strlen(const char *s);
+int								ft_strlen(const char *s);
 int									ft_isdigit(int c);
 char								*ft_strnstr(const char *big,
 										const char *little, size_t len);
